Fix the video software backend. (closes issue 4269) Some warning fixes and cleanup.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@7361 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/EmuWindow.cpp
@@ -199,7 +199,7 @@ LRESULT CALLBACK WndProc( HWND hWnd, UINT iMsg, WPARAM wParam, LPARAM lParam )
 		if (wParam == WM_USER_KEYDOWN)
 		{
 			OnKeyDown(lParam);
-			FreeLookInput(wParam, lParam);
+			FreeLookInput((u32)wParam, lParam);
 		}
 		else if (wParam == WIIMOTE_DISCONNECT)
 		{
